swaps epr & eprtest
